Update dependencies
GitHub complains that the TestNG version used
has a vulnerability, so while this is a pretty
much a dead tutorial project, it still isn't
a bad idea to fix it because otherwise someone
may clone it and use it as a base for... something.

While we're at it, ignore the .idea folder,
as it's a pure Maven project.